+++ CHANGELOG.md
@@ -6,10 +6,14 @@
 
 ### BUG FIXES
 
+- `[evidence]` fix flaky `TestReactorsGossipNoCommittedEvidence` test
+  ([\#5870](https://github.com/cometbft/cometbft/pull/5870))
 - `[blocksync]` fix removeTimedoutPeers deadlock found via Byzantine prevote gossip race
   ([\#5839](https://github.com/cometbft/cometbft/pull/5839))
 - `[mempool]` fix setRecheckFull/setDone race causing spurious ErrRecheckFull.
   ([\#5837](https://github.com/cometbft/cometbft/pull/5837))
+- `[node]` close partial listeners on startRPC failure
+  ([\#5869](https://github.com/cometbft/cometbft/pull/5869))
 
 ### IMPROVEMENTS